+++ src/twg2/dependency/jar/RepositoryInfo.java
@@ -16,12 +16,13 @@ import twg2.functions.IoFunc;
 /**
  * @author TeamworkGuy2
  * @since 2016-1-28
+ * @param <T> the type of data which can be read the projects in this repository
  */
-public interface RepositoryInfo {
+public interface RepositoryInfo<T> {
 
 	public String getName();
 
-	public RepositoryStructure getStructure();
+	public RepositoryStructure<T> getStructure();
 
 	public List<Path> getProjects();
 
@@ -31,19 +32,20 @@ public interface RepositoryInfo {
 	/**
 	 * @author TeamworkGuy2
 	 * @since 2016-1-28
+	 * @param <U> the type of data which can be read from the projects in the repository info instance being constructed by this builder
 	 */
-	public static class Builder {
+	public static class Builder<U> {
 		private static Path gitDir = Paths.get(".git");
 		private static Path hgDir = Paths.get(".hg");
 		private static Path svnDir = Paths.get(".svn");
 		private static DirectoryStream.Filter<Path> dirFilter = (p) -> !p.endsWith(gitDir) && !p.endsWith(hgDir) && !p.endsWith(svnDir);
 
 		private @Getter String name;
-		private @Getter RepositoryStructure structure;
+		private @Getter RepositoryStructure<U> structure;
 		private @Getter List<Path> projects;
 
 
-		public Builder(String name, RepositoryStructure structure) {
+		public Builder(String name, RepositoryStructure<U> structure) {
 			this.name = name;
 			this.structure = structure;
 			this.projects = new ArrayList<>();
@@ -85,8 +87,8 @@ public interface RepositoryInfo {
 		}
 
 
-		public RepositoryInfo.Impl build() {
-			return new RepositoryInfo.Impl(name, structure, new ArrayList<>(projects));
+		public RepositoryInfo.Impl<U> build() {
+			return new RepositoryInfo.Impl<>(name, structure, new ArrayList<>(projects));
 		}
 
 	}
@@ -97,14 +99,15 @@ public interface RepositoryInfo {
 	/**
 	 * @author TeamworkGuy2
 	 * @since 2016-1-28
+	 * @param <V> the type of data which can be read from the projects in this repository
 	 */
-	public static class Impl implements RepositoryInfo {
+	public static class Impl<V> implements RepositoryInfo<V> {
 		private final @Getter String name;
-		private final @Getter RepositoryStructure structure;
+		private final @Getter RepositoryStructure<V> structure;
 		private final @Getter List<Path> projects;
 
 
-		public Impl(String name, RepositoryStructure structure, List<Path> projects) {
+		public Impl(String name, RepositoryStructure<V> structure, List<Path> projects) {
 			this.name = name;
 			this.structure = structure;
 			this.projects = projects;